+++ lib/honeybadger/configuration.rb
@@ -7,8 +7,8 @@ module Honeybadger
                :ignore_user_agent, :notifier_name, :notifier_url, :notifier_version,
                :params_filters, :project_root, :port, :protocol, :proxy_host, :proxy_pass,
                :proxy_port, :proxy_user, :secure, :use_system_ssl_cert_chain, :framework,
-               :user_information, :rescue_rake_exceptions, :source_extract_radius,
-               :send_request_session, :debug, :fingerprint, :hostname].freeze
+               :user_information, :feedback, :rescue_rake_exceptions, :source_extract_radius,
+               :send_request_session, :debug, :fingerprint, :hostname, :metrics].freeze
 
     # The API key for your project, found on the project edit form.
     attr_accessor :api_key
@@ -84,7 +84,10 @@ module Honeybadger
     # The text that the placeholder is replaced with. {{error_id}} is the actual error number.
     attr_accessor :user_information
 
-    # The framework Honeybadger is configured to use
+    # Display user feedback form when configured?
+    attr_accessor :feedback
+
+    # The framework Honeybadger is configured to use.
     attr_accessor :framework
 
     # Should Honeybadger catch exceptions from Rake tasks?
@@ -173,6 +176,7 @@ module Honeybadger
       @metrics                   = true
       @features                  = { 'notices' => true }
       @limit                     = nil
+      @feedback                  = true
     end
 
     # Public: Takes a block and adds it to the list of backtrace filters. When