+++ spec/honeybadger/configuration_spec.rb
@@ -32,6 +32,7 @@ describe Honeybadger::Configuration do
     assert_config_default :debug, false
     assert_config_default :fingerprint, nil
     assert_config_default :hostname, Socket.gethostname
+    assert_config_default :feedback, true
   end
 
   it "configures async as Proc" do
@@ -113,6 +114,8 @@ describe Honeybadger::Configuration do
     assert_config_overridable :send_request_session
     assert_config_overridable :debug
     assert_config_overridable :hostname
+    assert_config_overridable :metrics
+    assert_config_overridable :feedback
   end
 
   it "has an api key" do
@@ -128,7 +131,7 @@ describe Honeybadger::Configuration do
      :notifier_version, :params_filters, :project_root, :port, :protocol,
      :proxy_host, :proxy_pass, :proxy_port, :proxy_user, :secure,
      :source_extract_radius, :async, :send_request_session, :debug,
-     :fingerprint, :hostname].each do |option|
+     :fingerprint, :hostname, :metrics, :feedback].each do |option|
        expect(hash[option]).to eq config[option]
     end
   end